[#50] Fix wrapping in footer
Put a Group around the Anchors in the footer so that they flex on one line, with space between.
Remove the h prop on the Footer in App.jsx, since that was causing some blank space if the footer content wasn't exactly that height.
Add missing prop-types and uuid packages.